Refactor CreateEvent method
