+++ Eventnet/Eventnet.Api.IntegrationTests/RabbitMqTestFactory.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using Eventnet.Api.Config;
 using Eventnet.Api.IntegrationTests.Helpers;
 using Eventnet.Api.IntegrationTests.Mocks;
 using Eventnet.Api.Services.SaveServices;
