+++ Eventnet/Eventnet.Api/Services/SaveServices/ConsumeEventService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Text;
 using System.Text.Json;
+using Eventnet.Api.Config;
 using Eventnet.Infrastructure;
 using RabbitMQ.Client;
 using RabbitMQ.Client.Events;
