+++ Eventnet/Eventnet.Infrastructure/Eventnet.Infrastructure.csproj
@@ -8,15 +8,15 @@
 
     <ItemGroup>
 
-        <PackageReference Include="AutoMapper" Version="11.0.1"/>
-        <PackageReference Include="System.Drawing.Common" Version="7.0.0-preview.2.22152.2"/>
+        <PackageReference Include="AutoMapper" Version="11.0.1" />
+        <PackageReference Include="System.Drawing.Common" Version="7.0.0-preview.2.22152.2" />
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Eventnet.DataAccess\Eventnet.DataAccess.csproj"/>
-        <ProjectReference Include="..\Eventnet.Domain\Eventnet.Domain.csproj"/>
+        <ProjectReference Include="..\Eventnet.DataAccess\Eventnet.DataAccess.csproj" />
+        <ProjectReference Include="..\Eventnet.Domain\Eventnet.Domain.csproj" />
 
-        <PackageReference Include="MailKit" Version="3.1.1"/>
+        <PackageReference Include="MailKit" Version="3.1.1" />
     </ItemGroup>
 
 </Project>
